+++ src/ConnectedTokenProvider.tsx
@@ -3,7 +3,7 @@ import { RootState } from 'src/redux/Store'
 import _isEmpty from 'lodash/isEmpty'
 import { useSelector } from 'react-redux'
 
-import { ThemeProvider } from '@mui/material'
+import { Theme, ThemeProvider } from '@mui/material'
 import { deepmerge } from '@mui/utils'
 
 import { createMXTheme } from '@kyper/mui'
@@ -31,7 +31,7 @@ interface ColorScheme {
   Brand500?: string
 }
 
-export const theme = (colorScheme: ColorScheme) => {
+export const clientThemeOverrides = (colorScheme: ColorScheme) => {
   if (!_isEmpty(colorScheme)) {
     return {
       palette: {
@@ -47,21 +47,45 @@ export const theme = (colorScheme: ColorScheme) => {
   }
   return {}
 }
+const connectThemeOverrides = (palette: Theme['palette']) => ({
+  components: {
+    MuiTypography: {
+      styleOverrides: {
+        root: {
+          // Sets the default color of text. Can override singluar usage with color prop.
+          color: palette?.text?.primary,
+        },
+      },
+    },
+  },
+})
+
 interface Props {
   children: React.ReactNode
 }
 
 export const ConnectedTokenProvider = ({ children }: Props): React.ReactNode => {
-  const customTokens = useSelector(getTokenProviderValues)
+  // Client custom tokens from batcave
+  const clientCustomTokens = useSelector(getTokenProviderValues)
+  // "Light" or "Dark"
   const colorScheme = useSelector((state: RootState) => state.config.color_scheme)
+  // Dark mode: true or false
   const isDarkModeEnabled: boolean = colorScheme === THEMES.DARK
+  // Theme object with MX overrides
   const mxTheme = createMXTheme(isDarkModeEnabled ? 'dark' : 'light')
-  const combinedTheme = deepmerge(mxTheme, theme(customTokens.tokenOverrides.Color))
+
+  const combinedTheme = deepmerge(
+    mxTheme,
+    deepmerge(
+      clientThemeOverrides(clientCustomTokens.tokenOverrides.Color), // Theme object with client custom overrides
+      connectThemeOverrides(mxTheme.palette as Theme['palette']), // Theme object with connect overrides
+    ),
+  )
 
   return (
     <TokenProvider
       theme={isDarkModeEnabled ? THEMES.DARK : colorScheme}
-      tokenOverrides={customTokens.tokenOverrides}
+      tokenOverrides={clientCustomTokens.tokenOverrides}
     >
       <ThemeProvider theme={combinedTheme}>{children}</ThemeProvider>
     </TokenProvider>